ffmpeg/vaapi: move VP8 encoder class to lib
