+++ lib/ffmpeg/vaapi/encoder.py
@@ -7,8 +7,9 @@
 import re
 import slash
 
+from ....lib import platform
 from ....lib.ffmpeg.encoderbase import BaseEncoderTest, Encoder as FFEncoder
-from ....lib.ffmpeg.util import have_ffmpeg_hwaccel
+from ....lib.ffmpeg.util import have_ffmpeg_hwaccel, have_ffmpeg_encoder
 from ....lib.ffmpeg.vaapi.util import mapprofile
 from ....lib.ffmpeg.vaapi.decoder import Decoder
 from ....lib.common import mapRangeInt
@@ -17,6 +18,12 @@ class Encoder(FFEncoder):
   hwaccel = property(lambda s: "vaapi")
   tilecols = property(lambda s: s.ifprop("tilecols", " -tile_cols {tilecols}"))
   tilerows = property(lambda s: s.ifprop("tilerows", " -tile_rows {tilerows}"))
+
+  @property
+  def profile(self):
+    if self.codec in ["vp8"]:
+      return ""
+    return super().profile
 
   @property
   def qp(self):
@@ -99,3 +106,25 @@ class EncoderTest(BaseEncoderTest):
     ]
     m = re.search(ipbmsgs[ipbmode], self.output, re.MULTILINE)
     assert m is not None, "Possible incorrect IPB mode used"
+
+############################
+## VP8 Encoders           ##
+############################
+
+@slash.requires(*have_ffmpeg_encoder("vp8_vaapi"))
+@slash.requires(*platform.have_caps("encode", "vp8"))
+class VP8EncoderTest(EncoderTest):
+  def before(self):
+    super().before()
+    vars(self).update(
+      codec     = "vp8",
+      ffenc     = "vp8_vaapi",
+      caps      = platform.get_caps("encode", "vp8"),
+      lowpower  = 0,
+    )
+
+  def get_file_ext(self):
+    return "ivf"
+
+  def get_vaapi_profile(self):
+    return "VAProfileVP8Version0_3"
